feat(UI): Add quinoa
Currently, the UI is built separately from the project.

This commit adds the quarkus quinoa extension to build the UI as part of
the project build.

How to use:
Launch the project in dev mode, and as part of the `wanaku-router`
model, the UI will be started in `dev` mode.

When building for production, the UI will be built and added to the
final artifact.

fix: https://github.com/wanaku-ai/wanaku/issues/76