fix(UI): Provide relative route for UI
When running the UI through quinoa, you will be accessing it through a quarkus endpoint, whereas when running the UI in standalone mode you will be accessing it from another server. This is important because depending on where you are running from, the API address will change.

The solution is to set a relative route in the `application.properties`
file so the services are call relative to the providing host.

For local development, a default value of `http://localhost:8080` is
provided.

fix: https://github.com/wanaku-ai/wanaku/issues/229