Abstract the downstream service registration implementation
Rework the service registration so it is done via the Router instead of services directly accessing the registry.

This greatly simplifies both implementing new services and the overall architecture of Wanaku as it moves most the configuration responsibilities to the service instance.

This solves issue #305 and helps with the implementation of #301.